ipmiconsole: validate numeric options before narrowing
Keep parsed payload-instance and proxy-port values as long integers
until their complete syntax and ranges have been checked, preventing
oversized inputs from wrapping into accepted values.

+++ ipmiconsole/ipmiconsole-argp.c
@@ -174,7 +174,7 @@ cmdline_parse (int key, char *arg, struct argp_state *state)
 {
   struct ipmiconsole_arguments *cmd_args;
   char *endptr;
-  int tmp;
+  long tmp;
 
   assert (state);
